Derive EPVBucket from VBucket class to implement persistence related logic
Currently we have EphemeralVBucket class that is derived from VBucket
class to implement logic related to only Ephemeral VBucket. Similarly we
also need a class to implement logic that is needed only for persistent
VBuckets.

VBucket class which is a parent class will implement logic that is common
to both Ephemeral and persistent VBuckets. Not planning to create just an
api only class (the one similar to KVBucketIface). VBucket class will serve
as both api and a class that implements common functionality.

                 ============
                 | VBucket  |
                 ============
                      |
                   ___ ___
                  |       |
        ===========      ==================
        |EPVBucket|      |EphemeralVBucket|
        ===========      ==================

Change-Id: Ic74e1bd13c9490380a6a52294ca93409e6a8a4cf
Reviewed-on: http://review.couchbase.org/73867
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>