MB-22747 [Ephemeral]: Update item count stats correctly after delete
In Ephemeral buckets we keep the deleted items in memory until they
are purged. We keep them in hash table to have O(1) access on deleted
items. This requires that we keep a count of deleted items in the
hash table so that we can keep other item count stats like curr_items
correct.

This commit adds 'numDeletedItems' stat in the hash table.

It also implements the VBucket::getNumItems() function separately in
the derived classes (persistant VB and ephemeral VB) so that we do the
counting in the Couchbase buckets the same way we do it now.

Finally it enables the test cases that were disabled specifically for
ephemeral buckets due to incorrect accounting of deleted items.

Change-Id: Ife9c15d33f84954aab5117ba3359326f061671e2
Reviewed-on: http://review.couchbase.org/74130
Reviewed-by: Oliver Downard <oliver.downard@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>