MB-22896: Add upgrade support for MutationLog.
5d97bde6 broke upgrades in that access log files are no longer
readable and will trigger asserts. The size of each MutationLogEntry
was increased by 1 byte and thus a access log from 4.6 will no longer
match codes assumptions.

To fix we now increase the version number of the MutationLog from 1
to 2. Next the code adds back the old MutationLogEntry structure so
that we can read a V1 file and convert the data to version 2.

The upgrade uses a serial upgrade technique where we will sequentially
upgrade from n to n + 1 and never skip versions. The intent is that
this will give simpler upgrade steps, but possibly higher cost if
there were many steps.

However in this code there can only be a V1 to V2 upgrade, but the
sturcture of the code is hopefully easily extensible for a few more
versions if we ever need to. After that a table driven approach maybe
better or something more exotic.

A test has been added which manually writes a V1 log file, that is
because MutationLog would need extensive work to make that fully
configurable to generate V1 files, so we just write the bytes into
a vector and flush to disk, then confirm we can now read the data
back.

Manual upgrade testing has also been performed.

Change-Id: Idceec80527ffecdccf2a077a6e00f5e7cb6762ce
Reviewed-on: http://review.couchbase.org/74156
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>