MB-22451: Revert the logging of ignoring request to notify stream ready
Change I75f2d25af5d0f11175beb7b23300486092aa86fa added logging when
ignoring a request to notify that a stream is ready due to the
associated vbucket already being in the ready queue.

Unfortunately this is unexpectantly flooding the logs therefore this
patch reverts the addition of this logging.

Change-Id: Id71ea83cb78637ab82b244719a2b4a12ea49a2e6
Reviewed-on: http://review.couchbase.org/74228
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>