In Ephemeral vb, bloomfilter to always return that key does not exist
In Ephemeral buckets, there is no need of bloomfilter. Until we
remove the bloomfilter in Ephemeral buckets, we make the bloomfilter
vb function which predicts if a key exists, to return false always.

By doing this we can avoid many Ephemeral Vb code paths that result in
bgFetchs. This commit sets stage for moving bgFetch functionality from
VBucket class to EPBucket class.

Change-Id: I85e296ccb9cc32aaebb9c0a26bc24dce81b3d2a2
Reviewed-on: http://review.couchbase.org/74288
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>