Eject/Evict hashtable items only in persistent Vbuckets
Item ejection (eviction), value only or full, is done only in
persistent vbuckets. In ephemeral vbuckets, for now there is no disk
to evict to. Hence these operations are not supported for now.

In future, we may decide to implement eviction as deletion.

By doing this, we also ensure that in Ephemeral vbuckets a non
resident item means a deleted item.

Change-Id: I56e0d041ce08aaaeadae1cfa015365dff0ee0330
Reviewed-on: http://review.couchbase.org/74290
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>