MB-22451: Backfill task leave stream state unchanged
The backfill task must not change the state machine of the active
stream that it is associated with.  In particular:
1) In ActiveStream::markDiskSnapshot if a vbucket is not found then
   it should just return, leaving the entity deleting the vbucket to
   set the stream to dead.
2) ActiveStream::completeBackfill should not call
   scheduleBackfill_UNLOCKED as this can move the stream into the
   STREAM_IN_MEMORY state.

State machine changes should only be driven by ActiveStream::next,
which is invoked by DcpProducer::getNextItem.

The call to scheduleBackfill_UNLOCKED that was invoked in
ActiveStream::completeBackfill when the pendingBackfill flag was true
has been moved to the ActiveStream::backfillPhase function and is
invoked once the current backfill has completed and the pendingBackfill
flag is set to true.

Change-Id: Idcbf8164792f4fd09898fe90748424687c60fb6a
Reviewed-on: http://review.couchbase.org/74388
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>