MB-22041 [9/N]: Revert thread limits to 512
A previous commit for this MB changed the limits to the number of each
type of worker thread to the number of CPU cores available. As this is
not optimal for synchronous, disk bound tasks returning the limit to the
previous 512 until a better limit is determined.

Additionally, this also led to the default values exceeding the value
which can be set in the config or by cbepctl. This would be unpleasant
to encounter after setting the number of threads down, as it would not
be possible to dynamically return to the original value.

A more significant issue introduced by the reduced limits - previously
valid configs setting a thread count greater than the number of CPU
cores are no longer accepted, breaking existing setups.

Change-Id: If8cfdc9e60f2bd36a37cf2ea6b378e47670581c7
Reviewed-on: http://review.couchbase.org/74402
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>