Move BG Fetch related logic to EPVBucket from VBucket
BG (background) Fetch is an operation needed only in persistant bucket
and mostly with full_eviction policy. Ephemeral buckets do not have
BG fetch hence do not support functions related to BG fetch.

This commit only moves the BG Fetch related logic in VBucket to
EPVBucket. There maybe other modules like 'KVBucket' class which will
be handled later.

In the base class VBucket, the eviction modes (VALUE and FULL) are not
abstracted very well. There are mostly implemented as if-else. Hence it was
not possible (easily) to make BG fetch operations private to the persitent
vbucket class EPVBucket. But the commit exploits the fact that BG fetch
operations are called only if FULL_EVICTION mode and in case of deleted
items that predicted to be on the disk by bloomfilter.

The commit assumes that Ephemeral vbuckets work only in "VALUE_EVICTION" policy
and that bloomfilter does not predict any item to be on the disk for Ephemeral
vbuckets. Hence logical exception is raised if any BG Fetch operation is called
in an Ephemeral vbucket.

Change-Id: I6d85b84a27395549bdfbf91abac857289cf156b3
Reviewed-on: http://review.couchbase.org/74067
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>