OrderedStoredValue [3/4]: Add OrderedStoredValue & factory
Add a subclass of StoredValue - OrderedStoredValue - which
additionally knows it's position in the seqno order. These objects are
constructed with a new OrderedStoredValueFactory (simiilar to
StoredValue), as they also have a variable-length key allocated inline
at the end of their fixed fields. Note that due to *not* having a
virtual destructor (instead we use flag dispatching), we need a custom
deleter for StoredValue::UniquePtr now.

See the StoredValue comment block for more details on
OrderedStoredValue's implementation.

Expand the StoredValue unit tests to also test this new class.

Change-Id: I2465fadee84d18a6ce449f5995ca34da82e2adb2
Reviewed-on: http://review.couchbase.org/74535
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>