MB-22960: Test to ensure no data loss with cursor dropping
There was a concern that if we have dropped the DCP cursor and the
persistence cursor moves forward by the time we re-register the DCP
cursor in the ActiveStream::markDiskSnapshot then we might lose data.

Therefore a test was created to investigate what happens in this
particular scenario.  The test showed that the endSeqno provided to
markDiskSnapshot is not the DCPBackfill::endSeqno but is instead the
last persisted seqno, which is provided by underlying storage engine.
This means there is not any data loss.

Change-Id: Ic23dc32ab3db596996e2e2a40a2d2c364009de3a
Reviewed-on: http://review.couchbase.org/74595
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>