MB-16181: Create/Begin using incorrect keys
The keys are wrong in our SystemEvents

The create event is correct:

A create event for collection "beer" stores

1. A special document in the system-namespace
   "$collections::create:beer<rev>"
2. Writes a new JSON manifest in _local.
3. Triggers DCP (it will do upstream)

The begin delete event is incorrect, it currently is set to delete
the special document written by the create, it is too soon to do that
as Items in the collection still exist.

The begin delete event needs to.
1. Trigger a DCP notification (tell a consumer X is deleting)
2. An update to the JSON manifest in _local.

The delete events (those that are created when we know we've purged
all items of a collection) are incorrect.

The delete events (hard/soft) need to:
1. Update the JSON manifest in the _local.
2. Delete the special document written by the create event.

Note the updating of documents and triggering of DCP is handled by
commits that add collection support in those areas, so this commit
only affects the SystemEvent keys.

Change-Id: Ibccd14e3da096b495c638303c4189bca0928d4c4
Reviewed-on: http://review.couchbase.org/74607
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>