MB-22018: Allowing varargs in cbstats commands
We used func.func_code.co_argcount (which may be considered peeking at
"internals") to check the number of arguments a function takes, to allow
us to print a nice exception rather than simply attempting to call it
with incorrect arguments and exiting from a TypeError.

Moving to using the `inspect` module, and checking if varargs are
accepted - allowing any number of arguments if so.

This may be used going forwards to delegate parsing of additional
options to a given command. This could be done currently but wouldn't be
very clean.

Change-Id: I8100f3e49c6b6143617493d267497c4b1ed6397b
Reviewed-on: http://review.couchbase.org/75427
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>