MB-22041 [10/N]: Add config aliasing
This will allow renaming of config values while maintaining backwards
compatability with, e.g., previously set extra_config_string values.

This is done by changing `attributes` in configuration.cc to use shared
pointers, and extending genconfig.cc to facilitate putting a second key
in the map with a pointer to the same value_t.

For example, in configuration.jason, to add an alias "cache_size" for
the parameter "max_size", one could do:

 "max_size": {
     "default": "0",
     "type": "size_t",
+    "aliases":["cache_size"]
 },

Additionally,

...
+    "aliases": "cache_size"
...

would be accepted in the case of a single alias.

Change-Id: Icd9e9fe0263f026b7e906b41efaea0e2b79ab676
Reviewed-on: http://review.couchbase.org/74694
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>