Extend engine api with "get_if" to optionally request documents
The primary motivation for this method is to allow the frontend
to optionally fetch items from the underlying engine if it
exists and meets certain criteras. This allows the underlying
engine to skip fetching the value for the items if the frontend
don't need the item after all.

This allows the frontend to check if a document exists and
contains any extended attributes without forcing the engine
to page the document back in from disk (if it don't contain
any attributes).

Change-Id: I3cb36a87a48cab9150e14bdf826489fd4195af39
Reviewed-on: http://review.couchbase.org/74759
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>