MB-23163: Fix log messsage in ActiveStream
The unexpected STREAM_DEAD log message has been seen in the logs
associated with the MB.

The log message was missing the vbucket that the stream is associated
with.  This has been added.

Changed to use producer->getLogger().log as opposed to the LOG macro,
as it matches other log messages in the file.  It ensures the
connection name is printed first.

Change-Id: Iade484be8d01f69aa5f9b8166af99a23285ff67e
Reviewed-on: http://review.couchbase.org/74766
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>