[Ephemeral Bucket]: Add a class for ordered sequence of items in memory
This commit adds an abstract base class 'SequenceList' that provides
apis to store ordered sequence of items (StoredValues) in memory in our
multi-threaded, monotonically increasing seq model.

It also adds 'BasicLinkedList', a concrete class that implements
SequenceList. 'BasicLinkedList' uses a boost intrusive doubly linked
list. It is essentially a linked list of 'StoredValues'.

Currently, only appendToList() functionality is implemented. But the
goal is, 'BasicLinkedList' works in tandem with already existing
'HashTable' to achieve O(1) writes/updates, O(1) reads and range reads.
Range read is necessary for an in-memory DCP backfill. Range reads
provide point-in-time snapshots which is essential for our incremental
DCP.

And a module test is added as well.

Change-Id: I9877c17c9e5b5be5bb55e559bbfb58f25281251d
Reviewed-on: http://review.couchbase.org/74440
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>