[Ephemeral Bucket]: Add point-in-time snapshot rangeRead on LinkedList
Monotonically point-in-time snapshots which can be used for
incremental DCP. This is implemented in the sequential data structure
'BasicLinkedList' in EphemeralVBuckets. This is equivalent to DCP disk
backfill in Couchbase buckets.

To get a point-in-time snapshot from the list, the main requirement is
no de-deduplication be allowed on the list in the range where the read
for the snapshot is happening. We achieve it by not blocking
the writes on the list during the time of the range read.

This commit implements only the range read part. The non-blocking
write during a range read will be evident with implementation of
updateListElem in an upcoming commit.

Change-Id: I80eeb760e94730645f4d20ec6a8de5ce81da7d26
Reviewed-on: http://review.couchbase.org/74639
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>