MB-22163: Use get_if to fetch items for mutations
The mutations _might_ need to preserve the xattrs from
the previous version of a document so we want to retrieve
only the documents which already contains xattrs.

The motivation for moving to get_if instead of get is
that the underlying engine may provide the metadata to
the callback without having the actual value for the
document in memory (whereas get() would require the
engine to fetch the document off disk). This would
result in a significant performance improvement in
DGM situations where we're running with value eviction.

For full eviction we must go to disk anyway.

Change-Id: I555320da1e2f200f67eb47264566ae92055816e7
Reviewed-on: http://review.couchbase.org/74842
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>