[Ephemeral]: Track memory usage of the items owned by the sequence list
In ephemeral buckets we can have stale items which are not owned by the
hash table, but owned by the sequence list. We need to track the
memory usage of these items in the list for overall memory tracking.

This commit only tracks memory used in the list. This is yet to be
exported out to all the necessary global stats.

Change-Id: Ie06e7a69a45780735a2d3f51341a261984c1786e
Reviewed-on: http://review.couchbase.org/74959
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>