Move Bucket construction to makeBucket()
To allow test code to construct the correct KVBucket subclass based on
the configuration, move the code to select and create the bucket from
EventuallyPersistentEngine::initialize to a seperate method -
makeBucket(). This will be used in subsequent Ephemeral bucket unit
tests.

As part of this, change EventuallyPersistentEngine::kvBucket from a
raw to a unique_ptr.

Change-Id: I527736f2f03a5eaa6effe6289deb15701f4c301b
Reviewed-on: http://review.couchbase.org/74993
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>