Add a method to copy an OrderedStoredValue
When we have to create a copy of an OrderedStoredValue in soft
deletes, we do not have the luxury of creating the copy from an
Item object. Hence we need a way to copy it from the existing
OrderedStoredValue object.

Soft deleted need a copy of OrderedStoredValue when the
OrderedStoredValue object being deleted needs to be around for the
concurrent point-in-time snapshot (backfill).

Change-Id: I5c7fc5e1d26cf05157b7563dc6f7e9fda9a52621
Reviewed-on: http://review.couchbase.org/74882
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>