[Ephemeral Bucket]: Put soft deletes on sequential storage
Since we are append-only, distributed cache supporting incremental
replication, we need to keep deleted items for while in the
sequential storage and periodically purge them.

Further, we also allow reads/writes on metadata of deleted items
and hence it becomes necessary to keep them in the HashTable.

Therefore, we treat soft deletes as updates on StoredValues and move
them to the correct position in the sequential storage. If there is
a range read on the sequential storage, then the corresponding
StoredValue is replaced in the HashTable with its copy and marked
as stale. The copy is then logically deleted and appended to the
sequential storage.

And along with the above, this commit tweaks how mark stale is called
and also updates the unit test module of the sequential storage.

Change-Id: I3ed3dabf9adf8ebddff5c602ebe1647ef1185ce4
Reviewed-on: http://review.couchbase.org/74897
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>