Dynamically create disk or in-memory (placeholder for now) backfill
Create the DCP backfill object dynamically in the vbucket object.
So if the vbucket object is of type persistant then disk backfill obj
is created and if it is of type ephemeral then in-memory backfill obj
can be created.

Currently we just have a placeholder for creating the backfill obj in
ephemeral vbuckets. It will be filled in once we have implementation
of in memory backfill in an upcoming commit.

Change-Id: I20f5fe1cc21e04a324adf6485c850d797a99ffd0
Reviewed-on: http://review.couchbase.org/75105
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>