Add unit tests for Item and Expiry pager
Add GoogleTest-style unit tests for the Item Pager and the Expiry
pager. Note these run both for Persistent and Ephemeral buckets.

Parts of the Item Pager test are currently disabled for Ephemeral, as
the Item Pager currently is hard-wired to evict items when paging them
out, and for Ephemeral this needs to be changed to a deletion.

Change-Id: I6083ccb05f4340808dad408a5a55fbadb640dec1
Reviewed-on: http://review.couchbase.org/75141
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>