HashTableVisitor: Include bucket lock in visit()
To allow HashTable visitor implementations to peform a larger range of
modifications to the item which is being visited (for example:
deleting the item), include a reference to the hash bucket lock in the
signature of HashTableVisitor::visit().

Also removed HashTableStatVisitor as it wasn't driven by the normal
visitor pattern and was simpler to inline the functionality than add
support for the bucket lock param.

Change-Id: I6b2d0e09be6d0044963f6ba5f00437d92b567f00
Reviewed-on: http://review.couchbase.org/75150
Reviewed-by: Oliver Downard <oliver.downard@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>