Create an abstract class DCPBackfill and concrete class DCPBackfillDisk
Currently DCP backfill is done only from the disk and is run by the
backfill manager. Going forward, for ephemeral buckets, we will want
the backfill to be done from an in memory data structure as well.

Hence this commit creates an abstract class for the backfill manager
to run the backfill and the concrete class DCPBackfillDisk to do the
backfill from the disk. Similarly we can extend the abstract class
to do in memory backfill in the future.

Also adds documentation (some, not all) for existing DCPBackfill(Disk),
removes some unused code.

Change-Id: Ia6f1d9bb25b6f48c7b8fac65cfc010c77816480d
Reviewed-on: http://review.couchbase.org/75097
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>