MB-16181: Integrate filters into DcpProducer and ActiveStream
The DCP producer now adds a filter to itself when opened, the filter
is configured from the DCP open input data.

Opening DCP without DCP_OPEN_COLLECTIONS allows access to the
default collection only (if it exists).

Opening DCP with DCP_OPEN_COLLECTIONS allows the client to specify
a filter or if no-value is given, a passthrough filter is created.

Change-Id: I6a8c24c0af20a07ddd029325487172b516e54b00
Reviewed-on: http://review.couchbase.org/78637
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>