AllocHooks: Switch to newer AllocHooks:: functions
Memory allocator hooks are exposed via a AllocHooks class, which was
added during the cbmalloc functionality. To assist in migration,
mc_XXX functions were kept around which called down to the equivient
mc_XXX function. Complete migration to the new AllocHooks::XXX static
functions, which will allow the compatibility functions to be deleted.

Change-Id: I237ce09ae9cf96ca99f5c337ae8e5f943aa3bd53
Reviewed-on: http://review.couchbase.org/75213
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>