MB-16181: Add getBySeqno to DcpResponse and return OptionalSeqno
This commit is useful up-stream where more DcpResponse sub-classes
now can return a seqno. This cleans up code where we check the event
type, then cast and pull out the seqno.

Upstream we would need more if's and a second cast to a different
response sub-class, but with this change we can just ask abstract
class for the seqno and check if it has a value.

Change-Id: Ib92390980ef394e94c6523131a5312b480bc8e44
Reviewed-on: http://review.couchbase.org/75396
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>