Ephemeral: Do DCP backfill from ordered in-memory data structure (list)
This commits implements a class that does backfill from in-memory
ordered data strucuture (list) and informs the DCP stream of the
backfill progress.

It calls a synchronous vBucket API to read items in the sequential
order from the in-memory ordered data strucuture and calls the DCP
stream for disk snapshot, backfill items and backfill completion.

This commit runs a DCP stream level module test to check if the
backfill happens correctly.

DCP functional tests for ephemeral buckets are not enabled in this
commit because they rely on certain stats and operations related
to persistence. Plan is to decouple these persistence dependencies
and enable them in a separate commit

Change-Id: Ia199602f96e62d44bd3f92fec7f8c9ad26a4dc83
Reviewed-on: http://review.couchbase.org/75184
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>