MB-22163: Use get_if to fetch items for mutations
The mutations _might_ need to preserve the xattrs from
the previous version of a document so we want to retrieve
only the documents which already contains xattrs.

The motivation for moving to get_if instead of get is
that the underlying engine may provide the metadata to
the callback without having the actual value for the
document in memory (whereas get() would require the
engine to fetch the document off disk). This would
result in a significant performance improvement in
DGM situations where we're running with value eviction.

For full eviction we must go to disk anyway.

This patch refactored the API call for get_if
to return a std::pair with the error code
in addition to the item as exception handling
cause a significant performance regression.

Change-Id: I00a2c268b5f1af327f3812b512a4ad66c5fd815f
Reviewed-on: http://review.couchbase.org/75486
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>