Allow env variables to tweak file logger
When running the server locally to reproduce it is nicer
to be able to do a "tail -f" on the memcached log and have
messages appear 'immediately' in the log file instead of
waiting up to 20 secs for it to be flushed. In these cases
it is also nice to be able to ignore file rotation.

Picking these variables up from environment is much easier
than having to figure out the correct magic to pass them
all the way down from ns_server ;)

Change-Id: Ic0785ccc754413018be8392cfeb2cf396dff225b
Reviewed-on: http://review.couchbase.org/75537
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>