MB-23067: Remove %n from printf
We use %n to count the number of characters printed by printf. This is a
secuirty hole and hence is disabled by default by MSVC. This patch
simply uses the length of the string passed rather than using %d. This
is to avoid doing some MSVC specific defs and because we try to avoid %n
in general.

Change-Id: I90f9bf117276516a3f15b31fef662ddce927d840
Reviewed-on: http://review.couchbase.org/75543
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>