Return high seqno as persisted seqno in ephemeral buckets
We do not have persistence in an ephemeral bucket. Hence we return
the last seen seqno (highSeqno) as the persisted seqno. This is needed
because higher layers like ns_server have long used persisted seqno as
last seen seqno for certain operations like vb takeover.

Also in ep-engine a "high priority vb event" is called during
vb-takeover to prioritize persistence. In ephemeral buckets we must
treat it as a no-op and return ENGINE_EWOULDBLOCK as returned in
regular couchbase buckets. This is because, an ephemeral vbucket must
wait for the highseqno to catch up to the desired seqno as well.

Change-Id: Iac79eaaa6efb4bb74472887c7de9a2dd854ce3be
Reviewed-on: http://review.couchbase.org/75587
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Haikney <david.haikney@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>