MB-23067: Explicitly make_pair in initializer_list
MSVC 2013 does not play well with the current initializer list for the
stats. This seems to be a variant of
https://connect.microsoft.com/VisualStudio/feedback/details/800104/

To accomodate for this bug, this commit explicitly uses make_pair and
specifies the types (the key ones being the std::strings) to help avoid
the compiler getting confused and silently changing the data.

One approach tried was to explicitly use an initializer_list, however
this still resulted in the same issue. Another approach of explicitly
creating std::strings in the initializer list resulted in an internal
compiler error.

Change-Id: I096c32503c7f5d1af8cad03c03d6372835242c3b
Reviewed-on: http://review.couchbase.org/75608
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>