[Ephemeral]: Update stream stats 'backfillRemaining' in memory backfills
'backfillRemaining' is a stat in DCP stream and can be used to track the
progress of a backfill. This has more significance in a disk backfill,
that is, how many items are yet to be read of the disk. In in-memory
backfill as well it is good to maintain this stat.

In in-memory backfills, we must increment it before reading the items to
the stream and decrement it as we read items one by one.

Change-Id: I9796d089fdb757b1c409c34a12df8bfd9269cf26
Reviewed-on: http://review.couchbase.org/75632
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>