MB-23505: [Ephemeral] Do not call KVStore file during vb-takeover stats
In Ephemeral buckets we do not have KVStore data file to get an estimate
of persisted deletes. We should rather get the estimate from in-memory
data structure.

We are already running vb-takeover tests on Ephemeral buckets in our
unit tests, but for some reason KVStore data file read does not fail
in unit tests. It only fails during a cluster run rebalance.
Change-Id: I5c1386d4358bae570d523dd50807f9acbe51f1bc
Reviewed-on: http://review.couchbase.org/75656
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>