MB-23505: Do not add items to vb backfill queue in Ephemeral buckets
Vb level backfill queue is for items in a huge snapshot (disk backfill
snapshots from DCP are typically huge) that could not be fit on a
checkpoint. They update all stats, checkpoint seqno, but are not put
on checkpoint and are directly persisted from the queue.

In ephemeral buckets we must not add backfill items from DCP (on
replica vbuckets), to the vb backfill queue because we have put them on
linkedlist already. Also we do not have the flusher task to drain the
items from that queue.
(Unlike checkpoints, the items in this queue is not cleaned up
 in a background cleanup task).

But we must be careful to update certain stats and checkpoint seqno
like in a regular couchbase bucket.

Further, this commit adds a stat 'backfill_queue_size' to track
the items in the vb backfill queue.

Change-Id: I8e513173a345d4c84d2e01778bad92d570d4754b
Reviewed-on: http://review.couchbase.org/75674
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>