MB-23522: Add CMD_{GET,SET} to multi-path ops
This commit modifies the binary protocol by allowing CMD_SETs and
CMD_GETs within multi-mutations and multi-lookups respectively. This is
to support both setting XATTRs and a whole body at the same time and
reading XATTRs and the whole doc at the same time.

These commands assume an empty path, if a non-empty path is used, EINVAL
will be returned.

The implementation of these commands largely make use of the already
existing infrastructure within subdocument. The key difference being
that they don't call down in to subjson to perform their operations.

Change-Id: Iea340a24bd924637b1f51644e95d3ab42dd90249
Reviewed-on: http://review.couchbase.org/75863
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>