MB-23505: Notify high priority rebalance requests in ephemeral buckets
High priority requests are async requests made by modules like ns-server
during rebalance. The request is for a response from the vbucket when it
'sees' beyond a certain sequence number or when a certain checkpoint is
persisted. Depending on the vbucket type, the meaning 'seeing' a sequence
number changes. That is, it could mean persisted in case of EPVBucket and
added to the sequenced data structure in case of EphemeralVBucket.

In ephemeral buckets we must notify high priority seqno requests as well.
Wwe notify when the requested seqno (or beyond) is put on the sequenced
data structure (say list) by the respective vbucket.

Change-Id: Ie2e255d762ac7fa3d5c2a6c817e62cf394b31406
Reviewed-on: http://review.couchbase.org/75716
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>