Modify access scanner memory benchmark
This change changes, slightly, what we are measuring with the benchmark.
We now record the maximum overhead per item actually stored in the
access scanner at any one time (the alog_max_stored_items) rather than
than the maximum overhead per item for all the items set during the
test. This is so we can more closely monitor the changes in memory
overhead for each item stored.

Change-Id: I8b5a67ca734fe59dc2f03a6e8c6b1fb11cd388e3
Reviewed-on: http://review.couchbase.org/75958
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>